feat : refreshAccessToken Lib 추가

+++ src/app/api/admin/kakao/unlink/route.js
@@ -1,10 +1,13 @@
+import userBanHandler from '@/app/(admin)/admin/_lib/userBanHandler';
+import { connectDB } from '@/lib/mongodb';
+import { ObjectId } from 'mongodb';
 import { NextResponse } from 'next/server';
 
 export async function POST(req) {
   try {
+    const client = await connectDB;
+    const db = client.db(process.env.MONGODB_NAME);
     const { providerAccountId, _id } = await req.json();
-
-    console.log(providerAccountId, _id);
 
     const body = new URLSearchParams({
       target_id_type: 'user_id',
@@ -25,6 +28,14 @@ export async function POST(req) {
       console.error(errorData);
       return NextResponse.json({ error: errorData }, { status: res.status });
     }
+
+    const { email } = await db.collection('users').findOne({ _id: new ObjectId(_id) });
+    const date = new Date();
+    date.setMonth(date.getMonth() + 3);
+    const expiresAt = Math.floor(date.getTime() / 1000);
+
+    const status = await userBanHandler(email, 0, expiresAt);
+    if (status !== 200) return NextResponse.json({}, { status });
 
     const res2 = await fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/api/user/${_id}`, { method: 'DELETE' });
     if (!res2.ok) {
